+++ CHANGELOG.md
@@ -21,6 +21,8 @@ If you like this project and find it useful, please consider giving it a star on
 - [frontend]: Use MbfTable for Plugins, Devices, Registered devices and Clusters tables.
 - [frontend]: Optimized WebSocker message handlers. Now, the handler targets the component.
 - [frontend]: Removed dangerouslySetInnerHTML from log rendering.
+- [endpoint]: Added occupancy feature to all Thermostat cluster helpers. When provided (either false or true) it will create a Thermostat with occupancy feature.
+- [endpoint]: Added outdoorTemperature to all Thermostat cluster helpers. Default is undefined (it will be ignored).
 
 ### Changed